modularize, turn VM into block machine, then enforce blocks
