+++ Cargo.toml
@@ -16,7 +16,8 @@ members = [
     "ast",
     "analysis",
     "linker",
-    "asm_utils"
+    "asm_utils",
+    "airgen",
 ]
 
 # Running "cargo build" on root directory will by default build just
