+++ airgen/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "asm_to_pil"
+name = "airgen"
 version = "0.1.0"
 edition = "2021"
