+++ analysis/src/lib.rs
@@ -1,4 +1,5 @@
 mod batcher;
+mod block_enforcer;
 mod function_id_setter;
 mod inference;
 mod macro_expansion;
@@ -11,14 +12,68 @@ pub use macro_expansion::MacroExpander;
 use ast::{asm_analysis::AnalysisASMFile, parsed::asm::ASMFile};
 use number::FieldElement;
 
+#[derive(Default)]
+struct DiffMonitor {
+    previous: Option<String>,
+    current: Option<String>,
+}
+
+impl DiffMonitor {
+    fn push<S: ToString>(&mut self, s: S) {
+        std::mem::swap(&mut self.previous, &mut self.current);
+        self.current = Some(s.to_string());
+
+        if let Some(current) = &self.current {
+            if let Some(previous) = &self.previous {
+                for diff in diff::lines(previous, current) {
+                    match diff {
+                        diff::Result::Left(l) => log::trace!("-{}", l),
+                        diff::Result::Both(..) => {}
+                        diff::Result::Right(r) => log::trace!("+{}", r),
+                    }
+                }
+            }
+        }
+    }
+}
+
 pub fn analyze<T: FieldElement>(file: ASMFile<T>) -> Result<AnalysisASMFile<T>, Vec<String>> {
-    let expanded = macro_expansion::expand(file);
-    let checked = type_check::check(expanded)?;
-    let inferred = inference::infer(checked)?;
-    let rommed = romgen::generate_rom(inferred);
-    let batched = batcher::batch(rommed);
-    let function_ids_set = function_id_setter::set(batched);
-    Ok(function_ids_set)
+    let mut monitor = DiffMonitor::default();
+
+    // expand macros
+    log::trace!("Run expand analysis step");
+    let file = macro_expansion::expand(file);
+    // type check
+    log::trace!("Run type-check analysis step");
+    let file = type_check::check(file)?;
+    monitor.push(&file);
+    // infer assignment registers (vm)
+    log::trace!("Run inference analysis step");
+    let file = inference::infer(file)?;
+    monitor.push(&file);
+    // generate the rom using a dispatcher (vm)
+    log::trace!("Run generate_rom analysis step");
+    let file = romgen::generate_rom(file);
+    monitor.push(&file);
+    // batch statements (vm)
+    log::trace!("Run batch analysis step");
+    let file = batcher::batch(file);
+    monitor.push(&file);
+    // generate the function ids from the rom (vm)
+    log::trace!("Run function_id_set analysis step");
+    let file = function_id_setter::set(file);
+    monitor.push(&file);
+    // turn asm into pil
+    log::trace!("Run asm_to_pil analysis step");
+    let file = asm_to_pil::compile(file);
+    monitor.push(&file);
+    // from now on we only have static machines!
+    // enforce blocks using function_id and latch
+    log::trace!("Run enforce_block analysis step");
+    let file = block_enforcer::enforce(file);
+    monitor.push(&file);
+
+    Ok(file)
 }
 
 pub mod utils {
