+++ analysis/src/romgen.rs
@@ -81,7 +81,7 @@
 use ast::{
     asm_analysis::{
         utils::previsit_expression_mut, AnalysisASMFile, AssignmentStatement, FunctionStatement,
-        InstructionStatement, Machine, PilBlock, Rom,
+        InstructionStatement, Machine, Rom,
     },
     parsed::{
         asm::{Param, ParamList},
@@ -92,8 +92,7 @@ use number::FieldElement;
 use std::{collections::HashMap, iter::once, marker::PhantomData, ops::ControlFlow};
 
 use crate::utils::{
-    parse_instruction_definition, parse_operation_statement, parse_pil_statement,
-    parse_register_declaration,
+    parse_instruction_definition, parse_operation_statement, parse_register_declaration,
 };
 
 /// Generate the ROM for each machine based on its functions
@@ -306,24 +305,6 @@ impl<T: FieldElement> RomGenerator<T> {
         } else {
             // do nothing for static machines
         };
-
-        let function_id = machine.function_id.as_ref().unwrap();
-        let latch = machine.latch.as_ref().unwrap();
-
-        // add the necessary embedded constraints which apply to both static and dynamic machines
-        let embedded_constraints = [
-            // inject the function_id
-            parse_pil_statement(&format!("pol witness {function_id}")),
-            // the function id must be constant within a block
-            parse_pil_statement(&format!(
-                "(1 - {latch}) * ({function_id}' - {function_id}) = 0"
-            )),
-        ];
-
-        machine.constraints.push(PilBlock {
-            start: 0,
-            statements: embedded_constraints.into_iter().collect(),
-        });
 
         machine
     }
