+++ compiler/Cargo.toml
@@ -23,6 +23,7 @@ json = "^0.12"
 ast = { version = "0.1.0", path = "../ast" }
 analysis = { version = "0.1.0", path = "../analysis" }
 linker = { version = "0.1.0", path = "../linker" }
+airgen = { version = "0.1.0", path = "../airgen" }
 
 [dev-dependencies]
 test-log = "0.2.12"
