+++ compiler/src/lib.rs
@@ -139,7 +139,7 @@ pub fn compile_asm_string<T: FieldElement>(
         panic!();
     });
     let analysed = analyze(parsed).unwrap();
-    let graph = asm_to_pil::compile(analysed);
+    let graph = airgen::compile(analysed);
     let pil = linker::link(graph);
 
     let pil_file_name = format!(
