+++ halo2/Cargo.toml
@@ -22,7 +22,7 @@ ast = { version = "0.1.0", path = "../ast" }
 analysis = { path = "../analysis" }
 executor = { path = "../executor" }
 parser = { path = "../parser" }
-asm_to_pil = { path = "../asm_to_pil" }
+airgen = { path = "../airgen" }
 test-log = "0.2.12"
 env_logger = "0.10.0"
 linker = { path = "../linker" }
