+++ linker/Cargo.toml
@@ -13,4 +13,4 @@ analysis = { path = "../analysis" }
 
 [dev-dependencies]
 parser = { path = "../parser" }
-asm_to_pil = { path = "../asm_to_pil" }
+airgen = { path = "../airgen" }
