+++ test_data/asm/vm_to_block_multiple_interfaces.asm
@@ -40,8 +40,8 @@ machine Main {
     }
 
     function main {
-        A <=Z= add(2, 1);
-        A <=Z= sub(A, 1);
+        A <== add(2, 1);
+        A <== sub(A, 1);
         assert_eq A, 2;
         loop;
     }
