+++ test_data/asm/vm_to_block_unique_interface.asm
@@ -64,11 +64,11 @@ machine Main {
     }
 
     function main {
-        A <=Z= add(2, 1);
-        A <=Z= sub(A, 1);
+        A <== add(2, 1);
+        A <== sub(A, 1);
         assert_eq A, 2;
-        A <=Z= and(1, 1);
-        A <=Z= or(A, 0);
+        A <== and(1, 1);
+        A <== or(A, 0);
         assert_eq A, 1;
         loop;
     }
